# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/video/generations": {
      "post": {
        "summary": "创建视频生成任务",
        "deprecated": false,
        "description": "提交视频生成任务，支持文生视频和图生视频。\n\n返回任务 ID，可通过 GET 接口查询任务状态。\n",
        "operationId": "createVideoGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "kling-v1",
                "prompt": "宇航员在月球上漫步",
                "duration": 5,
                "width": 1280,
                "height": 720
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建视频生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "错误信息"
              },
              "type": {
                "type": "string",
                "description": "错误类型"
              },
              "param": {
                "type": "string",
                "description": "相关参数",
                "nullable": true
              },
              "code": {
                "type": "string",
                "description": "错误代码",
                "nullable": true
              }
            }
          }
        }
      },
...
      "VideoRequest": {
        "type": "object",
        "description": "视频生成请求",
        "properties": {
          "model": {
            "type": "string",
            "description": "模型/风格 ID",
            "example": "kling-v1"
          },
          "prompt": {
            "type": "string",
            "description": "文本描述提示词",
            "example": "宇航员站起身走了"
          },
          "image": {
            "type": "string",
            "description": "图片输入 (URL 或 Base64)",
            "example": "https://example.com/image.jpg"
          },
          "duration": {
            "type": "number",
            "description": "视频时长（秒）",
            "example": 5
          },
          "width": {
            "type": "integer",
            "description": "视频宽度",
            "example": 1280
          },
          "height": {
            "type": "integer",
            "description": "视频高度",
            "example": 720
          },
          "fps": {
            "type": "integer",
            "description": "视频帧率",
            "example": 30
          },
          "seed": {
            "type": "integer",
            "description": "随机种子",
            "example": 20231234
          },
          "n": {
            "type": "integer",
            "description": "生成视频数量",
            "example": 1
          },
          "response_format": {
            "type": "string",
            "description": "响应格式",
            "example": "url"
          },
          "user": {
            "type": "string",
            "description": "用户标识",
            "example": "user-1234"
          },
          "metadata": {
            "type": "object",
            "description": "扩展参数 (如 negative_prompt, style, quality_level 等)",
            "additionalProperties": true,
            "properties": {}
          }
        }
      },
      "VideoResponse": {
        "type": "object",
        "description": "视频生成任务提交响应",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "任务 ID",
            "example": "abcd1234efgh"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "example": "queued"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session认证，通过登录接口获取"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Access Token认证，格式: Bearer {access_token}，通过 /api/user/token 接口生成"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "用户ID请求头，必须与当前登录用户ID匹配，使用Session或AccessToken认证时必须提供"
      },
      "Combination": {
        "group": [
          {
            "id": "SessionAuth"
          },
          {
            "id": "NewApiUser"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          },
          {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}